requirements.txt
Django~=4.2.2
django-environ~=0.10.0
django-cors-headers==3.14.*
djangorestframework==3.14.*
#drf-yasg[validation]
user_agents>=1.1.0
django-imagekit==4.1.*
Pillow~=10.0.0
#facebook-scraper
requests==2.31.*
django-admin-commands==1.0
ijson==3.2.*
django-jazzmin==2.6.*
django-related-admin==1.0.*
gunicorn==20.1.*
django-environ==0.10.*
django-import-export==3.2.*
psycopg~=3.1.9
#psycopg2cffi   #pypy not working.. meh..
#cx-Oracle~=8.3.0
pymemcache==4.0.*
django-picklefield==3.1
django-constance==2.9.*
dj-database-url==2.0.*
django-storages==1.13.*
boto3~=1.28.0
apify ~= 1.0.0
celery[redis]>=5.2.3,<6.0
django-celery-beat==2.5.*
django-celery-results==2.5.*
#django-password-policies   # TODO
flower
watchfiles~=0.19.0
django-cache-memoize~=0.1.10
SQLAlchemy~=2.0.21
